Completed version of High Fidelity Prototype
Digital (high-fidelity) prototype of a mobile triage app.

Triage is a process of sorting patients based on their need for immediate medical treatment according to certain guidelines. It enables prioritizing patients on the basis of the severity of their condition to receive medical facilities in a timely manner.
There are two main types of triage: i) hospital triage (SORT) which is generally conducted at a casualty clearing station or hospitals, and ii) field triage (SIEVE) which is performed out the hospital, e.g. in environments like mass gathering events where a large number of people gather, and it involves a rapid assessment. In this assignment, you will focus on SIEVE, i.e. field triage.

There are commonly three triage labels (R, Y, and G). These include:

1. the red label (R) indicates the immediate priority casualties (Priority 1),
2. the yellow/orange color (Y) which is used for those who require urgent treatment (Priority 2),
3. the green color (G) representing delayed priority causalities where patients have minor
injuries (Priority 3)

Using this app, first the user (most likely a paramedic or a nurse) needs to sign in with a user name and password (and to create an account if not exists). This process can be skipped by the user.
Then there will be five options to choose from: 

1) SEIVE triage,
2) Search by a patient, 
3) Search by a triage label (symptom or/and injuries), 
4) reports, and 
5) the map.

The triage labels are determined using simple rules based on the patient’s physiological data including their pulse, respiratory rate, and whether the patient is breathing and walking. The picture below shows how these parameters and rules are used to make a decision about the triage label. You do not need to include these rules in the application but the four parameters of pulse, respiratory rate, and whether the patient is breathing and walking should be included. Also, as the output, one of the triage labels should be displayed.